requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.3.0
pypdf>=4.0.0
pdfplumber>=0.11.0
streamlit>=1.38.0
requests>=2.32.0
tqdm>=4.66.0
python-dotenv>=1.0.0
